requirements.txt
psycopg2-binary==2.9.9
pillow==10.1.0
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
